requirements.txt
beautifulsoup4==4.4.1
click==5.1
lxml==3.5.0
pymongo==3.1
redis==2.10.5
requests==2.8.1
rq==0.5.6
wheel==0.24.0
tornado==4.3.0
python-logstash==0.4.5
raven==5.7.2
mongoengine==0.10.0
pep8==1.6.2
flake8==2.4.1
pyflakes==1.0.0
jedi==0.9.0
ipython==4.0.0
tox==2.1.1
pytest==2.8.2
PyYAML==3.11
